// src/lib/api/models/EventRevenueBreakdown.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EventRevenueBreakdown = {
    eventId?: string;
    eventTitle?: string;
    totalRevenue?: number;
    paidRevenue?: number;
    pendingRevenue?: number;
    totalRegistrations?: number;
    paidRegistrations?: number;
    pendingRegistrations?: number;
};